uhy_hiftuftuf

+++ src/main/java/utils/Etat.java
@@ -5,7 +5,7 @@ public class Etat {
     public final String nom;
     public final float duree;
 
-    public Etat(String nom, int duree) {
+    public Etat(String nom, float duree) {
         this.nom = nom;
         this.duree = duree;
     }
